requirements.txt
telethon-1.25.2